requirements.txt
pandas==1.3.4
streamlit==0.61.0
numpy==1.20.3
matplotlib==3.4.3
tensorflow==2.9.0
pillow==8.4.0
opencv-python==4.5.5.64
keras==2.9.0